packages moved, FPS printed in window title
