+++ src/main/java/fr/utbm/info/vi51/worldswar/gui/Window.java
@@ -4,7 +4,8 @@ import javax.swing.JFrame;
 import javax.swing.WindowConstants;
 
 import fr.utbm.info.vi51.worldswar.controller.Controller;
-import fr.utbm.info.vi51.worldswar.simulator.SimulationSpeed;
+import fr.utbm.info.vi51.worldswar.gui.messages.Messages;
+import fr.utbm.info.vi51.worldswar.gui.simulationgrid.GridViewPanel;
 import fr.utbm.info.vi51.worldswar.simulator.SimulatorListener;
 
 /**
@@ -14,22 +15,12 @@ import fr.utbm.info.vi51.worldswar.simulator.SimulatorListener;
  */
 public class Window extends JFrame implements SimulatorListener {
 	private static final long serialVersionUID = 3509021382819712013L;
-	
-	private Controller controller;
-	
+
+	private final Controller controller;
+
 	private float stepsPerSecond;
 	private long lastStepStart;
 
-	/**
-	 * @param args
-	 *            Temporary main method
-	 */
-	public static void main(String[] args) {
-		final Window window = new Window(null);
-		window.setVisible(true);
-	}
-
-	
 	/**
 	 * @param controller
 	 * 
@@ -38,11 +29,11 @@ public class Window extends JFrame implements SimulatorListener {
 	public Window(final Controller controller) {
 
 		// Controller initialization
-		
+
 		this.controller = controller;
 		this.stepsPerSecond = 0;
 		this.lastStepStart = System.currentTimeMillis();
-		
+
 		// Window initialization
 
 		this.setTitle(Messages.getString("Window.title")); //$NON-NLS-1$
@@ -57,14 +48,16 @@ public class Window extends JFrame implements SimulatorListener {
 		this.pack();
 		this.setVisible(true);
 	}
-	
+
 	@Override
 	public void simulationStepFired() {
-		long currentTime = System.currentTimeMillis();
-		long stepDuration = currentTime - this.lastStepStart;
+		final long currentTime = System.currentTimeMillis();
+		final long stepDuration = currentTime - this.lastStepStart;
 		this.lastStepStart = currentTime;
 		this.stepsPerSecond = 1000.f / stepDuration;
-		System.out.println(Messages.getString("Window.stepFiredMsg") + this.stepsPerSecond); //$NON-NLS-1$
+
+		final String title = String.format("%s - %f FPS", Messages.getString("Window.title"), this.stepsPerSecond);//$NON-NLS-1$//$NON-NLS-2$
+		this.setTitle(title);
 	}
 
 }
