+++ src/main/java/fr/utbm/info/vi51/worldswar/gui/messages/messages.properties
